Lavalink tests
